refactor: split settings into smaller components
